fix: refactor on singup for viewer
